Revert "[Impeller] support half precision uniforms and half precision samplers (#40590)"
This reverts commit 3287e2b362fcae6e755ed0022f02175cb5a368bb.

+++ impeller/entity/shaders/solid_fill.frag
@@ -5,11 +5,11 @@
 #include <impeller/types.glsl>
 
 uniform FragInfo {
-  f16vec4 color;
+  vec4 color;
 }
 frag_info;
 
-out f16vec4 frag_color;
+out vec4 frag_color;
 
 void main() {
   frag_color = frag_info.color;
